# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Personal Expense Tracker API
  version: 1.0.0
  description: Local REST API for the Personal Expense Tracker
  license:
    name: MIT
servers:
- url: http://localhost:3000
  description: Local backend
components:
  schemas:
    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum:
          - expense
          - income
        date:
          type: string
          format: date
        amountCents:
          type: integer
        categoryId:
          type: integer
          nullable: true
        incomeSourceId:
          type: integer
          nullable: true
        paymentMethodId:
          type: integer
        note:
          type: string
          maxLength: 500
          nullable: true
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        attachmentId:
          type: string
          nullable: true
        recurringRuleId:
          type: integer
          nullable: true
        occurrenceDate:
          type: string
          format: date
          nullable: true
      required:
      - id
      - type
      - date
      - amountCents
      - paymentMethodId
...
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
          nullable: true
        sortOrder:
          type: integer
        archived:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
      - id
      - name
      - sortOrder
      - archived
...
    WishListItem:
      type: object
      properties:
        id:
          type: integer
        month:
          type: integer
        year:
          type: integer
        name:
          type: string
        estimatedCostCents:
          type: integer
        categoryId:
          type: integer
        priority:
          type: string
          enum:
          - high
          - medium
          - low
        note:
          type: string
          maxLength: 500
          nullable: true
        status:
          type: string
          enum:
          - planned
          - purchased
          - abandoned
        linkedTransactionId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
      - id
      - month
      - year
      - name
      - estimatedCostCents
      - categoryId
      - priority
      - status
...
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
      required:
      - error
paths:
...
  /api/transactions:
    get:
      summary: List transactions
      parameters:
      - name: dateFrom
        in: query
        schema:
          type: string
          format: date
      - name: dateTo
        in: query
        schema:
          type: string
          format: date
      - name: type
        in: query
        schema:
          type: string
      - name: categoryId
        in: query
        schema:
          type: integer
      - name: incomeSourceId
        in: query
        schema:
          type: integer
      - name: paymentMethodId
        in: query
        schema:
          type: integer
      - name: tagId
        in: query
        schema:
          type: integer
      - name: page
        in: query
        schema:
          type: integer
          default: 1
      - name: pageSize
        in: query
        schema:
          type: integer
          default: 50
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  total:
                    type: integer
                required:
                - items
                - total
        '400':
          description: Bad Request
        '428':
          description: Precondition Required
      operationId: getTransactions
    post:
      summary: Create transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum:
                  - expense
                  - income
                date:
                  type: string
                  format: date
                amountCents:
                  type: integer
                categoryId:
                  type: integer
                incomeSourceId:
                  type: integer
                paymentMethodId:
                  type: integer
                note:
                  type: string
                tagIds:
                  type: array
                  items:
                    type: integer
              required:
              - type
              - date
              - amountCents
              - paymentMethodId
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '428':
          description: Precondition Required
      operationId: postTransactions
  /api/transactions/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
    get:
      summary: Get transaction
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '404':
          description: Not Found
        '428':
          description: Precondition Required
      operationId: getTransactionsId
    put:
      summary: Update transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum:
                  - expense
                  - income
                date:
                  type: string
                  format: date
                amountCents:
                  type: integer
                categoryId:
                  type: integer
                incomeSourceId:
                  type: integer
                paymentMethodId:
                  type: integer
                note:
                  type: string
                tagIds:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '428':
          description: Precondition Required
      operationId: putTransactionsId
    delete:
      summary: Delete transaction
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
        '428':
          description: Precondition Required
      operationId: deleteTransactionsId
...
  /api/wishlist:
    get:
      summary: List wish list items
      parameters:
      - name: year
        in: query
        required: true
        schema:
          type: integer
      - name: month
        in: query
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/WishListItem'
                  estimatedTotalCents:
                    type: integer
        '400':
          description: Bad Request
        '428':
          description: Precondition Required
      operationId: getWishlist
    post:
      summary: Create wish list item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                month:
                  type: integer
                year:
                  type: integer
                name:
                  type: string
                estimatedCostCents:
                  type: integer
                categoryId:
                  type: integer
                priority:
                  type: string
                  enum:
                  - high
                  - medium
                  - low
                note:
                  type: string
              required:
              - month
              - year
              - name
              - estimatedCostCents
              - categoryId
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WishListItem'
        '400':
          description: Bad Request
        '428':
          description: Precondition Required
      operationId: postWishlist
  /api/wishlist/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
    put:
      summary: Update wish list item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                estimatedCostCents:
                  type: integer
                categoryId:
                  type: integer
                priority:
                  type: string
                  enum:
                  - high
                  - medium
                  - low
                note:
                  type: string
                status:
                  type: string
                  enum:
                  - planned
                  - abandoned
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WishListItem'
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '428':
          description: Precondition Required
      operationId: putWishlistId
    delete:
      summary: Delete wish list item
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
        '428':
          description: Precondition Required
      operationId: deleteWishlistId
  /api/wishlist/{id}/purchase:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
    post:
      summary: Convert to transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amountCents:
                  type: integer
                date:
                  type: string
                  format: date
                paymentMethodId:
                  type: integer
              required:
              - paymentMethodId
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  wishListItem:
                    $ref: '#/components/schemas/WishListItem'
                  transaction:
                    $ref: '#/components/schemas/Transaction'
        '400':
          description: Bad Request
        '409':
          description: Conflict
        '428':
          description: Precondition Required
      operationId: postWishlistIdPurchase
  /api/wishlist/{id}/copy-next-month:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
    post:
      summary: Carry over
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WishListItem'
        '404':
          description: Not Found
        '428':
          description: Precondition Required
      operationId: postWishlistIdCopy-next-month
...
security: []